+++ src/main/java/com/example/demo/Doctor/DoctorService.java
@@ -5,10 +5,10 @@ import java.util.List;
 public interface DoctorService {
     List<Doctor> obtenerTodos();
     void agregar(Doctor doctor);
-    Doctor buscarPorId(int id);
+    Doctor buscarPorId(Long id);
     Doctor buscarDoctor(String dni);
     void actualizar(Doctor doctor);
-    void eliminar(int id);
+    void eliminar(Long id);
     List<Doctor> buscarPorDni(String dni);
